Remove Qubit from c++ library and move it to pure Python
