+++ CMakeLists.txt
@@ -216,7 +216,7 @@ add_library(netket MODULE Sources/pynetket.cc Sources/Dynamics/py_dynamics.cpp
     Sources/Graph/hypercube.cc Sources/Graph/lattice.cc
     Sources/Graph/custom_graph.cc Sources/Graph/abstract_graph.cc
     Sources/Graph/py_graph.cc Sources/Hilbert/hilbert_index.cc
-    Sources/Hilbert/bosons.cc Sources/Hilbert/spins.cc Sources/Hilbert/qubits.cc
+    Sources/Hilbert/bosons.cc Sources/Hilbert/spins.cc 
     Sources/Hilbert/custom_hilbert.cc Sources/Hilbert/py_hilbert.cc
     Sources/Machine/abstract_machine.cc Sources/Machine/rbm_spin.cc
     Sources/Machine/rbm_multival.cc Sources/Machine/rbm_spin_phase.cc
